refactor: move registry/registry to registry pkg
