+++ cmd/manifest/init-v2.go
@@ -37,7 +37,7 @@ import (
 	oktetoLog "github.com/okteto/okteto/pkg/log"
 	"github.com/okteto/okteto/pkg/model"
 	"github.com/okteto/okteto/pkg/okteto"
-	"github.com/okteto/okteto/pkg/registry/registry"
+	"github.com/okteto/okteto/pkg/registry"
 	"github.com/spf13/afero"
 	"github.com/spf13/cobra"
 )
