+++ internal/test/build.go
@@ -16,18 +16,17 @@ package test
 import (
 	"context"
 
-	oktetoErrors "github.com/okteto/okteto/pkg/errors"
 	"github.com/okteto/okteto/pkg/types"
 )
 
 // FakeOktetoBuilder emulates an okteto image builder
 type FakeOktetoBuilder struct {
 	Err      []error
-	Registry *FakeOktetoRegistry
+	Registry *fake.FakeOktetoRegistry
 }
 
 // NewFakeOktetoBuilder creates a FakeOktetoBuilder
-func NewFakeOktetoBuilder(registry *FakeOktetoRegistry, errors ...error) *FakeOktetoBuilder {
+func NewFakeOktetoBuilder(registry *fake.FakeOktetoRegistry, errors ...error) *FakeOktetoBuilder {
 	return &FakeOktetoBuilder{
 		Err:      errors,
 		Registry: registry,
@@ -48,49 +47,4 @@ func (fb *FakeOktetoBuilder) Run(_ context.Context, opts *types.BuildOptions) er
 		}
 	}
 	return nil
-}
-
-// FakeOktetoRegistry emulates an okteto image registry
-type FakeOktetoRegistry struct {
-	Err      error
-	Registry map[string]*FakeImage
-}
-
-// FakeImage represents the data from an image
-type FakeImage struct {
-	Registry string
-	Repo     string
-	Tag      string
-	ImageRef string
-	Args     []string
-}
-
-// NewFakeOktetoRegistry creates a new registry if not already created
-func NewFakeOktetoRegistry(err error) *FakeOktetoRegistry {
-	return &FakeOktetoRegistry{
-		Err:      err,
-		Registry: map[string]*FakeImage{},
-	}
-}
-
-// AddImageByName adds an image to the registry
-func (fb *FakeOktetoRegistry) AddImageByName(images ...string) error {
-	for _, image := range images {
-		fb.Registry[image] = &FakeImage{}
-	}
-	return nil
-}
-
-// AddImageByOpts adds an image to the registry
-func (fb *FakeOktetoRegistry) AddImageByOpts(opts *types.BuildOptions) error {
-	fb.Registry[opts.Tag] = &FakeImage{Args: opts.BuildArgs}
-	return nil
-}
-
-// GetImageTagWithDigest returns the image tag digest
-func (fb *FakeOktetoRegistry) GetImageTagWithDigest(imageTag string) (string, error) {
-	if _, ok := fb.Registry[imageTag]; !ok {
-		return "", oktetoErrors.ErrNotFound
-	}
-	return imageTag, nil
 }
