+++ pkg/registry/image.go
