--- pkg/registry/registry/fake/registry.go
@@ -1,1 +1,0 @@
-package fake
