+++ pkg/registry/replacer.go
