+++ pkg/registry/replacer_test.go
